backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "litellm>=1.83.9",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.44.0",
]
